Split up export svg test
