+++ eslint.config.mjs
@@ -180,7 +180,6 @@ export default tseslint.config(
         },
       ],
 
-      '@typescript-eslint/no-deprecated': 'off',
       '@typescript-eslint/no-explicit-any': 'off',
       '@typescript-eslint/explicit-module-boundary-types': 'off',
       '@typescript-eslint/ban-ts-comment': 'off',
