+++ products/jbrowse-web/src/tests/ExportSvgBreakpointSplitView.test.tsx
@@ -32,10 +32,12 @@ test('export svg of breakpoint split view', async () => {
   fireEvent.click(await findByText('Submit'))
 
   await waitFor(() => {
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
     expect(FileSaver.saveAs).toHaveBeenCalled()
   }, delay)
 
   // @ts-expect-error
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
   const svg = FileSaver.saveAs.mock.calls[0][0].content[0]
   const dir = path.dirname(module.filename)
   fs.writeFileSync(
